`cf logs` uses v7
Unlike most commands, the `cf logs` doesn't hit many CAPI endpoints;
rather, it hits the log endpoints, which can be found on the root, (i.e.
GET /). Here is a redacted version of the logging links. Note that the
`logging` link (doppler) is deprecated, and instead we use the newer
`log_cache` interfaces:

```json
{
   "links": {
      "logging": {
         "href": "wss://doppler.iris-eater.lite.cli.fun:443"
      },
      "log_cache": {
         "href": "https://log-cache.iris-eater.lite.cli.fun"
      }
   }
}
```

[#168654370]

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>